Cleaning Model and adding configs
Model now uses callbacks to provide the progress bar and show metrics. We also return a dictionary of outputs now to make it easier to know what each metric is.
Config files are now added and contain all possible arguments the modules could need. In theory, the configs can be significantly shorter if defaults are appropriate.