New pyproject.toml for Hatch
Using pyproject.toml allows us to use Hatch for setting up environments to run the code repeatably/reliably. Currently the only dependency that is version controlled is Keras 3.4.1 due to a warning I was seeing with 3.6. We should probably be more strict about versioning since we haven't done testing to know what works and what doesn't.